Began writing logic for page switching and resuming
