+++ src/index.css
@@ -106,7 +106,7 @@ p {
   }
 }
 
-#canvas-container {
+.canvas-container {
   position: absolute;
   top: 0;
   width: 100%;
